Initialize file pointer to NULL
The original code depended on unconditionally running the filename
resolution code, but for UKL we skip this for the linked application.
This leaves the file pointer intialized with what ever garbage was on
the stack previously. Initialize this pointer to NULL so that the UKL
application doesn't trip on chasing garbage pointers.

Signed-off-by: Eric B Munson <munsoner@bu.edu>