x86/fault: Skip checking kernel mode access to user address space for UKL
Normally, this check ensures that a kernel task has not ended up somehow
raising a page fault in the user part of address space. This is done by
checking if the CS value on stack. UKL always has the kernel value so this
check will always fail. This change makes sure that this check is only done
for non-UKL tasks by checking the in_user flag.

Signed-off-by: Ali Raza <aliraza@bu.edu>