Add UKL Bypass Mode
The UKL Base model allows for the UKL process to interact with the
kernel by making a direct function call to the system call handler.
However, the system call handler does a number of book keeping jobs on
top of selecting the appropriate system call and routing the results
back such as RCU management and signal and scheduler checking. We may
not want these to happen on each invocation from the UKL process.

This patch introduces two related changes: the ability to control the
bypass of entry/exit code and symbols to be used to do the bypass. These
changes give the UKL process the ability to toggle bypass at runtime and
to control how often the standard entry/exit code is run. By default we
route every 11th call through the standard entry and exit to ensure that
the work done there still happens. The new symbols are generated as part
of the system call definition macros and are callable directly by the
UKL process.

Signed-off-by: Ali Raza <aliraza@bu.edu>