Allow UKL process to use RET instead of IRET
The IRET instruction is more expensive than RET and while syscalls are
now using RET instead of SYSRET, interrupts and faults still use the
more expensive exit when returning to the UKL process. Add a config
option that allows the kernel to use RET instead of IRET when returning
to the UKL process from an interrupt or fault.

Signed-off-by: Ali Raza <aliraza@bu.edu>