Make copy_from_user() UKL aware
copy_from_user() checks the input address and the size to make sure that
the entire copied region have valid, user space addresses. When the
calling task is UKL, this may not be the case (e.g. the source address
is in UKL .data or .bss segments). The hacky solution is to accept the
full 64bits as valid source addresses for the UKL tasks.

Signed-off-by: Eric B. Munson <munsoner@bu.edu>