Add timestamp collection for event latency
We want to capture the time between when the kernel is aware that a
notification needs to be sent to the UKL process and when that process
finally handles the event. In the redis case, this is right before redis
calls read(). Take the timestamp when we know that we are waking the UKL
process.

Signed-off-by: Eric B. Munson <munsoner@bu.edu>