Event handler execution
The core idea here is that when we want to handle an event for the
target task (which is a UKL task), we only need its mm_struct. The text
for the UKL task is linked to the kernel and so already mapped, we just
need the page tables for the task. So when we register an event handler
we make note of the task and we will borrow its mm_struct when we go to
execute the event handler.

Signed-off-by: Eric B. Munson <munsoner@bu.edu>