Disable interrupts on UKL syscall entry
The SYSCALL instruction is setup to disable interrupts as part of the
ring transition. Because we use CALL instead to get to the entry point
for a UKL process, we come in with interrupts enabled and need to
explicitly disable them first thing. This also means that they must be
explicitly enabled on the way out.

Signed-off-by: Eric B Munson <munsoner@bu.edu>