Check if data is already available at registration
It is possible for data to arrive on a connected socket before the
handler is ready to consume it, so we need to check if this is the case.
When data already exists, the vfs_poll() function will return a bitfield
of events which have already occurred, if this bit field contains
EPOLLIN | EPOLLRDNORM then data is aready available and a work item
needs to be inserted into the queue (and a worker woken to handle).

Signed-off-by: Eric B Munson <munsoner@bu.edu>