Make event handler per-cpu
The will allow us to avoid generating IPI wawke ups when we schedule the
event handler task and try to make better use of locality. It also
allows us to drop the locking around the event queue as we can guaratee
that access to the queue by both producer and consumer are atomic when
interrupts are disabled.

Signed-off-by: Eric B Munson <munsoner@bu.edu>