Move setting task state to idle above work check
We want our worker task to be prepared to be woken up by a call to
ttwu() as soon as possible. When we get into worker_sleep() we have
already checked the work queue once and found it empty. On our way to
sleep, we mark ourselves idle and check again.

Signed-off-by: Eric B Munson <munsoner@bu.edu>