Remove references to modules so debugging works with MODULES=n
All of the GDB scripts make references to to modules and the module code
in configured headers. However, if a kernel is built with MODULES=n, as
UKL is, this code will fail and prevent the loading of the other GDB
helper functions. There is probably a cleaner way that uses the config
values than this hammer, but modules are not going to be a thing for
UKL for now.

Signed-off-by: Eric B Munson <munsoner@bu.edu>