Reorder wait queue management and ensure we check for existing events
We want to ensure that there is no possible race between wait queue
management and event handling so ensure that we remove the wait queue
entry before we enqueue a new event and make sure we reattach to the
wait queue before we call the handler. This leaves a small window for a
spurious event in the queue but cannot miss any.

Signed-off-by: Eric B Munson <munsoner@bu.edu>