Remove task list addition at worker registration
We unconditionally add the worker to the event handler task list before
the worker goes to sleep. However, if we have already added the worker
to the list at registration, this will corrupt the list. We can simply
skip adding at registration becuase the next time this worker is idle it
will add itself to that list.

Signed-off-by: Eric B Munson <munsoner@bu.edu>