Double check the work queue before sleeping
Prior to this commit our sleep function would race with event arrivals
if the worker was preempted after checking for new work but before
sleeping (as these steps were not atomic). This commit adds a second
check of the work queue once we hold the task lock. This ensures that if
data arrived in the race window, we catch it. It also ensures that if
data arrives after this point, the wake up will succeed.

Signed-off-by: Eric B Munson <munsoner@bu.edu>