Overhaul the upcall interface
We have a reasonable wrapper for user space, but the kernel system was
held together by duct tape and good vibes. Replace that with something
more robust. We manage subscriptions to event streams in a way that
mimics epoll, but without the added complexity from being able to nest
epoll file descriptors inside eachother.

Signed-off-by: Eric B Munson <munsoner@bu.edu>