Move memory barrier inside lock and remove sched rt for workers
The correct place for the memory barrier is pior to releasing the lock
to ensure that any task waiting on the lock can see all writes to the
protected structures.

Experiments show that the sched rt with very high priority did not
change latency, throughput, or scalability in an appreciable way. Remove
this configuration and use sched_other like everything else.

Signed-off-by: Eric B Munson <munsoner@bu.edu>